Create pot.py
The teacher has sent an e-mail to her students with the following task: “Write a program that will determine and output the value of XX if given the statement:

X=numberpow11+numberpow22+…+numberpowNN
X=number1pow1+number2pow2+…+numberNpowN
and it holds that number1number1, number2number2 to numberNnumberN are integers, and pow1pow1, pow2pow2 to powNpowN are one-digit integers.” Unfortunately, when the teacher downloaded the task to her computer, the text formatting was lost so the task transformed into a sum of NN integers:

X=P1+P2+…+PN
X=P1+P2+…+PN
For example, without text formatting, the original task in the form of X=212+1253X=212+1253 became a task in the form of X=212+1253X=212+1253. Help the teacher by writing a program that will, for given NN integers from P1P1 to PNPN determine and output the value of XX from the original task.