Create faktor.py
The impact factor of a scientific journal is a measure reflecting the average number of citations to articles published in science journals. For this task we are using a simplified formula for calculating the impact factor:

Total count of citations received by articles published in the journalTotal number of articles published
Totalcount of citations received by articles published in thejournalTotal number of articles published
Rounding is always performed upwards. For example the impact factor of the “Journal for ore research and time wasting” that published 3838 articles cited 894894 times is 894/38=23.53894/38=23.53 rounding up to 2424.

You are the editor of one scientific journal. You know how many articles you are going to publish and the owners are pushing you to reach a specific impact factor. You are wondering how many scientists you will have to bribe to cite your articles to meet the owners demands. Since money is tight you want to bribe the minimal amount of scientists. Each bribed scientist buys you a single citation.